+++ CHANGELOG.md
@@ -12,3 +12,8 @@ All notable changes to `dependencies` will be documented in this file
 ## 0.1.1 - 2021-07-08
 - add url & response contains assertions
 - add packages to the dataProvider
+
+
+## 0.2.0 - 2021-07-09
+- refactor test suite to use `Http` facade instead of `GuzzleHttp\Client`
+- refactor `DependencyService` methods to return an instance of `DependencyUrl` for retrieving urls & svgs